Tools: Moved ArrowSymbols extension.
